Make fields import-able from webargs.fields
See discussion on https://github.com/sloria/webargs/issues/61#issuecomment-142166317
