+++ webargs/__init__.py
@@ -1,11 +1,21 @@
 # -*- coding: utf-8 -*-
-
 from marshmallow.utils import missing
-from webargs.core import Nested, WebargsError, ValidationError
+# Make marshmallow's validation functions importable from webargs
+from marshmallow import validate
+
+from webargs.core import argmap2schema, WebargsError, ValidationError
+from webargs import fields
 
 __version__ = '0.15.0'
 __author__ = 'Steven Loria'
 __license__ = 'MIT'
 
 
-__all__ = ['WebargsError', 'ValidationError', 'missing', 'Nested']
+__all__ = (
+    'argmap2schema',
+    'WebargsError',
+    'ValidationError',
+    'fields',
+    'missing',
+    'validate',
+)
